sui_open_rpc_macros

Attribute Macro open_rpc

#[open_rpc]

Add a [Service name]OpenRpc struct and implementation providing access to Open RPC doc builder. This proc macro must be use in conjunction with jsonrpsee_proc_macro::rpc

The generated method open_rpc is added to [Service name]OpenRpc, ideally we want to add this to the trait generated by jsonrpsee framework, creating a new struct to provide access to the method is a workaround.

TODO: consider contributing the open rpc doc macro to jsonrpsee to simplify the logics.